Define CodeMirror in external codeMirror.ts
